requirements.txt
requests==2.31.0
PyYAML==6.0.1
flask==3.1.0
pytest==8.3.5
pytest-mock==3.14.0
